+++ src/main/java/sk/freemap/gpxAnimator/TileCache.java
@@ -61,14 +61,16 @@ public class TileCache {
     //
     // If either check fails, log a warning rather than delete the file.
     //
-    public static void ageCache(final String tileCachePath, final Long tileCacheTimeLimit) {
-        if (cachingEnabled(tileCachePath)) {
+    public static void ageCache() {
+        final String tileCacheDir = Preferences.getTileCacheDir();
+        final long tileCacheTimeLimit = Preferences.getTileCacheTimeLimit();
+        if (cachingEnabled(tileCacheDir)) {
             // Remove any cached tiles that are too old
-            File cacheDir = new File(tileCachePath);
+            final File cacheDir = new File(tileCacheDir);
             final File[] files = cacheDir.listFiles();
             if (files != null) {
                 for (File cacheEntry : files) {
-                    String cacheFilename = cacheEntry.getName();
+                    final String cacheFilename = cacheEntry.getName();
                     if ((cacheFilename.length() == 74) && (cacheFilename.endsWith(cachedFileExtension))) {
                         ageCacheFile(cacheEntry, tileCacheTimeLimit);
                     } else {
@@ -79,13 +81,13 @@ public class TileCache {
         }
     }
 
-    public static BufferedImage getTile(final String url, final String tileCachePath, final Long tileCacheTimeLimit) throws UserException {
+    public static BufferedImage getTile(final String url, final String tileCacheDir, final Long tileCacheTimeLimit) throws UserException {
 
         BufferedImage image;
 
-        if (cachingEnabled(tileCachePath)) {
+        if (cachingEnabled(tileCacheDir)) {
             try {
-                image = cachedGetTile(url, tileCachePath, tileCacheTimeLimit);
+                image = cachedGetTile(url, tileCacheDir, tileCacheTimeLimit);
             } catch (final UserException e) {
                 image = unCachedGetTile(url);
             }
@@ -111,10 +113,10 @@ public class TileCache {
         return mapTile;
     }
 
-    private static BufferedImage cachedGetTile(final String url, final String tileCachePath, final Long tileCacheTimeLimit) throws UserException {
+    private static BufferedImage cachedGetTile(final String url, final String tileCacheDir, final Long tileCacheTimeLimit) throws UserException {
         BufferedImage mapTile = null;
         String filename = hashName(url) + cachedFileExtension;
-        String path = tileCachePath + File.separator + filename;
+        String path = tileCacheDir + File.separator + filename;
         File cacheFile = new File(path);
 
         // Age out old tile file in cache directory.